+++ changelog.md
@@ -1,5 +1,6 @@
 # v0.3.0
 1. asgi: 2.0 -> 3.0
+2. StaticFiles: dir => dir/index.html, not found(html=True): 404.html
 
 
 # v0.2.5